se cambian a long todos los ID, y se ajusta el update de alumno

+++ src/main/java/com/bew/demo/dao/DocsBajaRepository.java
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.bew.demo.model.DocsBaja;
 
-public interface DocsBajaRepository extends JpaRepository<DocsBaja, Integer> {
+public interface DocsBajaRepository extends JpaRepository<DocsBaja, Long> {
 	
 	@Query(
             value = "SELECT s FROM DocsBaja s WHERE s.idDoc = :idDoc",
