+++ src/main/java/com/bew/demo/dao/SolicitudBajaRepository.java
@@ -12,12 +12,12 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 @Repository
-public interface SolicitudBajaRepository extends JpaRepository <SolicitudBaja,Integer>{
+public interface SolicitudBajaRepository extends JpaRepository <SolicitudBaja,Long>{
 	
 	@Query(
             value = "SELECT s FROM SolicitudBaja s WHERE s.idAlumno = :idAlumno",
             nativeQuery = false)
-    Optional<SolicitudBaja> findByIdAlumno(@Param("idAlumno") Integer idAlumno);
+    Optional<SolicitudBaja> findByIdAlumno(@Param("idAlumno") Long idAlumno);
 	
 	@Query(
             value = "SELECT s FROM SolicitudBaja s WHERE s.estado = :estado",
