+++ src/main/java/com/bew/demo/service/ClienteService.java
@@ -9,10 +9,10 @@ import com.bew.demo.exception.EmptyResultException;
 public interface ClienteService {
 
 	List<ClienteDTO> findAll();
-	ClienteDTO findById(Integer idCliente);
+	ClienteDTO findById(Long idCliente);
 	void saveCliente(ClienteDTO clienteDTO);
     void updateCliente(ClienteDTO clienteDTO)throws EmptyResultException;
-    void deleteCliente(Integer idCliente) throws EmptyResultException;
+    void deleteCliente(Long idCliente) throws EmptyResultException;
    // ClienteDTO ClienteNombre (String nombre) throws EmptyResultException ;
     //Optional <ClienteDTO> clientesNombre (String nombre) throws EmptyResultException ;
 	List<ClienteDTO> clientesNombre(String nombre);
