+++ src/main/java/com/bew/demo/service/DocsServicioService.java
@@ -8,7 +8,7 @@ import com.bew.demo.exception.EmptyResultException;
 
 public interface DocsServicioService {
 
-	ResponseEntity<ByteArrayResource> load(Integer fileId) throws EmptyResultException;
+	ResponseEntity<ByteArrayResource> load(Long fileId) throws EmptyResultException;
 	String store(MultipartFile file, String idDoc) throws EmptyResultException;
 	ResponseEntity<ByteArrayResource> findDoc(String idDoc);
 	void deleteDoc (String idDoc);
